config: Add config option for a `deployments.json` file
