+++ util/src/lib.rs
@@ -10,6 +10,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 pub mod logging;
 pub mod networking;
 pub mod runtime;
+pub mod starknet;
 
 /// Returns the current unix timestamp in seconds, represented as u64
 pub fn get_current_time_seconds() -> u64 {
